Fixed memory type

+++ src/main/scala/wvlet/querybase/interval/IntervalLike.scala
@@ -9,7 +9,7 @@ trait IntervalLike[A]:
     def updateWith(created_time: Long, start_time: Long, finished_time: Long): A
     def end: Long
     def cpuTime: Long
-    def memoryTime: Long
+    def memoryTime: Double
     def length: Long                                   = end - start
     def contains[B: IntervalLike](other: B): Boolean   = a.start <= other.start && other.end <= a.end
     def contains(pos: Long): Boolean                   = a.start <= pos && pos < a.end
